Define ignorable? on VVA errors
**Why**: So that Efolder can use the same pattern as Caseflow when
determining when to not send errors to Sentry.

For now, we assume all VVA errors should go to Sentry. If we find there
are some we can ignore, we can define them like we do in the BGS and
VBMS gems.
